Implemented changes based on PR comments
