+++ src/modules/block/components/blocks/blocks.js
@@ -7,7 +7,7 @@ import BoxContent from 'src/theme/box/content';
 import FilterBar from 'src/modules/common/components/filterBar';
 import StickyHeader from 'src/theme/table/stickyHeader';
 import Table from 'src/theme/table';
-import { transformStringDateToUnixTimestamp } from 'src/utils/dateTime';
+import transformParams from 'src/utils/transformParams';
 import { useBlocks } from '../../hooks/queries/useBlocks';
 import BlocksOverview from '../blocksOverview';
 import BlockFilterDropdown from './blockFilterDropdown';
@@ -42,35 +42,6 @@ const Blocks = ({
     /* istanbul ignore next */
     generatorAddress: (value) => `${t('Generated by')}: ${value}`,
   };
-
-  const transformParams = (params) =>
-    Object.keys(params).filter((param) => params[param] !== '').reduce((acc, item) => {
-      switch (item) {
-        case 'dateFrom':
-          if (params[item]) {
-            if (!acc.timestamp) acc.timestamp = ':';
-            acc.timestamp = acc.timestamp.replace(
-              /(\d+)?:/,
-              `${transformStringDateToUnixTimestamp(params[item])}:`,
-            );
-          }
-          break;
-        case 'dateTo':
-          if (params[item]) {
-            if (!acc.timestamp) acc.timestamp = ':';
-            // We add 86400 so the range is inclusive
-            acc.timestamp = acc.timestamp.replace(
-              /:(\d+)?/,
-              `:${transformStringDateToUnixTimestamp(params[item]) + 86400}`,
-            );
-          }
-          break;
-        default:
-          acc[item] = params[item];
-      }
-
-      return acc;
-    }, {});
 
   const handleLoadMore = () => {
     const searchParams = Object.keys(filters).reduce(
